errors resolved
